move `hash-ontology-icon` to `shared` folder
