+++ packages/hash/frontend/src/pages/[account-slug]/shared/hash-ontology-icon.tsx
