+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/[entity-type-id].page.tsx
@@ -20,7 +20,7 @@ import {
   EntityTypeEditorForm,
   EntityTypeEditorPropertyData,
 } from "./form-types";
-import { HashOntologyIcon } from "./hash-ontology-icon";
+import { HashOntologyIcon } from "../../shared/hash-ontology-icon";
 import { OntologyChip } from "../../shared/ontology-chip";
 import { PropertyListCard } from "./property-list-card";
 import { useEntityType } from "./use-entity-type";
